requirements.txt
# FastAPI Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Google Gemini AI (Free API)
google-generativeai>=0.3.2

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Market Data (Free)
yfinance>=0.2.36
nsepython

# Web Scraping
beautifulsoup4>=4.12.3
requests>=2.31.0

# Database
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Testing
pytest>=7.4.0
httpx>=0.26.0

# AI/ML Models
darts>=0.27.0
transformers>=4.36.0
torch>=2.1.0
scikit-learn>=1.3.0
prophet>=1.1.5
